LPS-79336 Update message board UAD implementation with renames
