+++ modules/apps/collaboration/message-boards/message-boards-uad-test/src/testIntegration/java/com/liferay/message/boards/uad/anonymizer/test/MBCategoryUADAnonymizerTest.java
@@ -16,10 +16,10 @@ package com.liferay.message.boards.uad.anonymizer.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 
-import com.liferay.message.boards.model.MBMessage;
-import com.liferay.message.boards.service.MBMessageLocalService;
+import com.liferay.message.boards.model.MBCategory;
+import com.liferay.message.boards.service.MBCategoryLocalService;
 import com.liferay.message.boards.uad.constants.MBUADConstants;
-import com.liferay.message.boards.uad.test.MBMessageUADEntityTestHelper;
+import com.liferay.message.boards.uad.test.MBCategoryUADEntityTestHelper;
 
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -28,8 +28,8 @@ import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
 import com.liferay.user.associated.data.aggregator.UADAggregator;
-import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
-import com.liferay.user.associated.data.test.util.BaseUADEntityAnonymizerTestCase;
+import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
+import com.liferay.user.associated.data.test.util.BaseUADAnonymizerTestCase;
 import com.liferay.user.associated.data.test.util.WhenHasStatusByUserIdField;
 
 import org.junit.After;
@@ -46,39 +46,38 @@ import java.util.List;
  * @generated
  */
 @RunWith(Arquillian.class)
-public class MBMessageUADEntityAnonymizerTest
-	extends BaseUADEntityAnonymizerTestCase<MBMessage>
+public class MBCategoryUADAnonymizerTest extends BaseUADAnonymizerTestCase<MBCategory>
 	implements WhenHasStatusByUserIdField {
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule = new LiferayIntegrationTestRule();
 
 	@Override
-	public MBMessage addBaseModelWithStatusByUserId(long userId,
+	public MBCategory addBaseModelWithStatusByUserId(long userId,
 		long statusByUserId) throws Exception {
-		MBMessage mbMessage = _mbMessageUADEntityTestHelper.addMBMessageWithStatusByUserId(userId,
+		MBCategory mbCategory = _mbCategoryUADEntityTestHelper.addMBCategoryWithStatusByUserId(userId,
 				statusByUserId);
 
-		_mbMessages.add(mbMessage);
+		_mbCategories.add(mbCategory);
 
-		return mbMessage;
+		return mbCategory;
 	}
 
 	@Override
-	protected MBMessage addBaseModel(long userId) throws Exception {
+	protected MBCategory addBaseModel(long userId) throws Exception {
 		return addBaseModel(userId, true);
 	}
 
 	@Override
-	protected MBMessage addBaseModel(long userId, boolean deleteAfterTestRun)
+	protected MBCategory addBaseModel(long userId, boolean deleteAfterTestRun)
 		throws Exception {
-		MBMessage mbMessage = _mbMessageUADEntityTestHelper.addMBMessage(userId);
+		MBCategory mbCategory = _mbCategoryUADEntityTestHelper.addMBCategory(userId);
 
 		if (deleteAfterTestRun) {
-			_mbMessages.add(mbMessage);
+			_mbCategories.add(mbCategory);
 		}
 
-		return mbMessage;
+		return mbCategory;
 	}
 
 	@Override
@@ -87,21 +86,21 @@ public class MBMessageUADEntityAnonymizerTest
 	}
 
 	@Override
-	protected UADEntityAnonymizer getUADEntityAnonymizer() {
-		return _uadEntityAnonymizer;
+	protected UADAnonymizer getUADAnonymizer() {
+		return _uadAnonymizer;
 	}
 
 	@Override
 	protected boolean isBaseModelAutoAnonymized(long baseModelPK, User user)
 		throws Exception {
-		MBMessage mbMessage = _mbMessageLocalService.getMBMessage(baseModelPK);
+		MBCategory mbCategory = _mbCategoryLocalService.getMBCategory(baseModelPK);
 
-		String userName = mbMessage.getUserName();
-		String statusByUserName = mbMessage.getStatusByUserName();
+		String userName = mbCategory.getUserName();
+		String statusByUserName = mbCategory.getStatusByUserName();
 
-		if ((mbMessage.getUserId() != user.getUserId()) &&
+		if ((mbCategory.getUserId() != user.getUserId()) &&
 				!userName.equals(user.getFullName()) &&
-				(mbMessage.getStatusByUserId() != user.getUserId()) &&
+				(mbCategory.getStatusByUserId() != user.getUserId()) &&
 				!statusByUserName.equals(user.getFullName())) {
 			return true;
 		}
@@ -111,7 +110,7 @@ public class MBMessageUADEntityAnonymizerTest
 
 	@Override
 	protected boolean isBaseModelDeleted(long baseModelPK) {
-		if (_mbMessageLocalService.fetchMBMessage(baseModelPK) == null) {
+		if (_mbCategoryLocalService.fetchMBCategory(baseModelPK) == null) {
 			return true;
 		}
 
@@ -119,26 +118,26 @@ public class MBMessageUADEntityAnonymizerTest
 	}
 
 	@Override
-	public void tearDownBaseModels(List<MBMessage> baseModels)
+	public void tearDownBaseModels(List<MBCategory> baseModels)
 		throws Exception {
-		_mbMessageUADEntityTestHelper.cleanUpDependencies(baseModels);
+		_mbCategoryUADEntityTestHelper.cleanUpDependencies(baseModels);
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		_mbMessageUADEntityTestHelper.cleanUpDependencies(_mbMessages);
+		_mbCategoryUADEntityTestHelper.cleanUpDependencies(_mbCategories);
 	}
 
 	@DeleteAfterTestRun
-	private final List<MBMessage> _mbMessages = new ArrayList<MBMessage>();
+	private final List<MBCategory> _mbCategories = new ArrayList<MBCategory>();
 	@Inject
-	private MBMessageLocalService _mbMessageLocalService;
+	private MBCategoryLocalService _mbCategoryLocalService;
 	@Inject
-	private MBMessageUADEntityTestHelper _mbMessageUADEntityTestHelper;
+	private MBCategoryUADEntityTestHelper _mbCategoryUADEntityTestHelper;
 	@Inject(filter = "model.class.name=" +
-	MBUADConstants.CLASS_NAME_MB_MESSAGE)
+	MBUADConstants.CLASS_NAME_MB_CATEGORY)
 	private UADAggregator _uadAggregator;
 	@Inject(filter = "model.class.name=" +
-	MBUADConstants.CLASS_NAME_MB_MESSAGE)
-	private UADEntityAnonymizer _uadEntityAnonymizer;
+	MBUADConstants.CLASS_NAME_MB_CATEGORY)
+	private UADAnonymizer _uadAnonymizer;
 }
