+++ modules/apps/collaboration/message-boards/message-boards-uad-test/src/testIntegration/java/com/liferay/message/boards/uad/anonymizer/test/MBThreadUADAnonymizerTest.java
@@ -16,10 +16,10 @@ package com.liferay.message.boards.uad.anonymizer.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 
-import com.liferay.message.boards.model.MBCategory;
-import com.liferay.message.boards.service.MBCategoryLocalService;
+import com.liferay.message.boards.model.MBThread;
+import com.liferay.message.boards.service.MBThreadLocalService;
 import com.liferay.message.boards.uad.constants.MBUADConstants;
-import com.liferay.message.boards.uad.test.MBCategoryUADEntityTestHelper;
+import com.liferay.message.boards.uad.test.MBThreadUADEntityTestHelper;
 
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -28,8 +28,8 @@ import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
 import com.liferay.user.associated.data.aggregator.UADAggregator;
-import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
-import com.liferay.user.associated.data.test.util.BaseUADEntityAnonymizerTestCase;
+import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
+import com.liferay.user.associated.data.test.util.BaseUADAnonymizerTestCase;
 import com.liferay.user.associated.data.test.util.WhenHasStatusByUserIdField;
 
 import org.junit.After;
@@ -46,39 +46,38 @@ import java.util.List;
  * @generated
  */
 @RunWith(Arquillian.class)
-public class MBCategoryUADEntityAnonymizerTest
-	extends BaseUADEntityAnonymizerTestCase<MBCategory>
+public class MBThreadUADAnonymizerTest extends BaseUADAnonymizerTestCase<MBThread>
 	implements WhenHasStatusByUserIdField {
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule = new LiferayIntegrationTestRule();
 
 	@Override
-	public MBCategory addBaseModelWithStatusByUserId(long userId,
+	public MBThread addBaseModelWithStatusByUserId(long userId,
 		long statusByUserId) throws Exception {
-		MBCategory mbCategory = _mbCategoryUADEntityTestHelper.addMBCategoryWithStatusByUserId(userId,
+		MBThread mbThread = _mbThreadUADEntityTestHelper.addMBThreadWithStatusByUserId(userId,
 				statusByUserId);
 
-		_mbCategories.add(mbCategory);
+		_mbThreads.add(mbThread);
 
-		return mbCategory;
+		return mbThread;
 	}
 
 	@Override
-	protected MBCategory addBaseModel(long userId) throws Exception {
+	protected MBThread addBaseModel(long userId) throws Exception {
 		return addBaseModel(userId, true);
 	}
 
 	@Override
-	protected MBCategory addBaseModel(long userId, boolean deleteAfterTestRun)
+	protected MBThread addBaseModel(long userId, boolean deleteAfterTestRun)
 		throws Exception {
-		MBCategory mbCategory = _mbCategoryUADEntityTestHelper.addMBCategory(userId);
+		MBThread mbThread = _mbThreadUADEntityTestHelper.addMBThread(userId);
 
 		if (deleteAfterTestRun) {
-			_mbCategories.add(mbCategory);
+			_mbThreads.add(mbThread);
 		}
 
-		return mbCategory;
+		return mbThread;
 	}
 
 	@Override
@@ -87,21 +86,23 @@ public class MBCategoryUADEntityAnonymizerTest
 	}
 
 	@Override
-	protected UADEntityAnonymizer getUADEntityAnonymizer() {
-		return _uadEntityAnonymizer;
+	protected UADAnonymizer getUADAnonymizer() {
+		return _uadAnonymizer;
 	}
 
 	@Override
 	protected boolean isBaseModelAutoAnonymized(long baseModelPK, User user)
 		throws Exception {
-		MBCategory mbCategory = _mbCategoryLocalService.getMBCategory(baseModelPK);
+		MBThread mbThread = _mbThreadLocalService.getMBThread(baseModelPK);
 
-		String userName = mbCategory.getUserName();
-		String statusByUserName = mbCategory.getStatusByUserName();
+		String userName = mbThread.getUserName();
+		String statusByUserName = mbThread.getStatusByUserName();
 
-		if ((mbCategory.getUserId() != user.getUserId()) &&
+		if ((mbThread.getUserId() != user.getUserId()) &&
 				!userName.equals(user.getFullName()) &&
-				(mbCategory.getStatusByUserId() != user.getUserId()) &&
+				(mbThread.getRootMessageUserId() != user.getUserId()) &&
+				(mbThread.getLastPostByUserId() != user.getUserId()) &&
+				(mbThread.getStatusByUserId() != user.getUserId()) &&
 				!statusByUserName.equals(user.getFullName())) {
 			return true;
 		}
@@ -111,7 +112,7 @@ public class MBCategoryUADEntityAnonymizerTest
 
 	@Override
 	protected boolean isBaseModelDeleted(long baseModelPK) {
-		if (_mbCategoryLocalService.fetchMBCategory(baseModelPK) == null) {
+		if (_mbThreadLocalService.fetchMBThread(baseModelPK) == null) {
 			return true;
 		}
 
@@ -119,26 +120,24 @@ public class MBCategoryUADEntityAnonymizerTest
 	}
 
 	@Override
-	public void tearDownBaseModels(List<MBCategory> baseModels)
+	public void tearDownBaseModels(List<MBThread> baseModels)
 		throws Exception {
-		_mbCategoryUADEntityTestHelper.cleanUpDependencies(baseModels);
+		_mbThreadUADEntityTestHelper.cleanUpDependencies(baseModels);
 	}
 
 	@After
 	public void tearDown() throws Exception {
-		_mbCategoryUADEntityTestHelper.cleanUpDependencies(_mbCategories);
+		_mbThreadUADEntityTestHelper.cleanUpDependencies(_mbThreads);
 	}
 
 	@DeleteAfterTestRun
-	private final List<MBCategory> _mbCategories = new ArrayList<MBCategory>();
+	private final List<MBThread> _mbThreads = new ArrayList<MBThread>();
 	@Inject
-	private MBCategoryLocalService _mbCategoryLocalService;
+	private MBThreadLocalService _mbThreadLocalService;
 	@Inject
-	private MBCategoryUADEntityTestHelper _mbCategoryUADEntityTestHelper;
-	@Inject(filter = "model.class.name=" +
-	MBUADConstants.CLASS_NAME_MB_CATEGORY)
+	private MBThreadUADEntityTestHelper _mbThreadUADEntityTestHelper;
+	@Inject(filter = "model.class.name=" + MBUADConstants.CLASS_NAME_MB_THREAD)
 	private UADAggregator _uadAggregator;
-	@Inject(filter = "model.class.name=" +
-	MBUADConstants.CLASS_NAME_MB_CATEGORY)
-	private UADEntityAnonymizer _uadEntityAnonymizer;
+	@Inject(filter = "model.class.name=" + MBUADConstants.CLASS_NAME_MB_THREAD)
+	private UADAnonymizer _uadAnonymizer;
 }
