+++ modules/apps/collaboration/message-boards/message-boards-uad/src/main/java/com/liferay/message/boards/uad/anonymizer/MBThreadUADAnonymizer.java
@@ -22,8 +22,8 @@ import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.User;
 
-import com.liferay.user.associated.data.anonymizer.DynamicQueryUADEntityAnonymizer;
-import com.liferay.user.associated.data.anonymizer.UADEntityAnonymizer;
+import com.liferay.user.associated.data.anonymizer.DynamicQueryUADAnonymizer;
+import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
 import com.liferay.user.associated.data.util.UADAnonymizerHelper;
 
 import org.osgi.service.component.annotations.Component;
@@ -37,8 +37,8 @@ import java.util.List;
  * @generated
  */
 @Component(immediate = true, property =  {
-	"model.class.name=" + MBUADConstants.CLASS_NAME_MB_THREAD}, service = UADEntityAnonymizer.class)
-public class MBThreadUADEntityAnonymizer extends DynamicQueryUADEntityAnonymizer<MBThread> {
+	"model.class.name=" + MBUADConstants.CLASS_NAME_MB_THREAD}, service = UADAnonymizer.class)
+public class MBThreadUADAnonymizer extends DynamicQueryUADAnonymizer<MBThread> {
 	@Override
 	public void autoAnonymize(MBThread mbThread, long userId)
 		throws PortalException {
