+++ modules/apps/collaboration/message-boards/message-boards-uad/src/main/java/com/liferay/message/boards/uad/exporter/MBMessageUADExporter.java
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.message.boards.uad.exporter;
+
+import com.liferay.message.boards.model.MBMessage;
+import com.liferay.message.boards.service.MBMessageLocalService;
+import com.liferay.message.boards.uad.constants.MBUADConstants;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+
+import com.liferay.user.associated.data.exporter.DynamicQueryUADExporter;
+import com.liferay.user.associated.data.exporter.UADExporter;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Brian Wing Shun Chan
+ * @generated
+ */
+@Component(immediate = true, property =  {
+	"model.class.name=" + MBUADConstants.CLASS_NAME_MB_MESSAGE}, service = UADExporter.class)
+public class MBMessageUADExporter extends DynamicQueryUADExporter<MBMessage> {
+	@Override
+	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
+		return _mbMessageLocalService.getActionableDynamicQuery();
+	}
+
+	@Override
+	protected String[] doGetUserIdFieldNames() {
+		return MBUADConstants.USER_ID_FIELD_NAMES_MB_MESSAGE;
+	}
+
+	@Reference
+	private MBMessageLocalService _mbMessageLocalService;
+}
